% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anagram-package.R
\docType{package}
\name{anagram-package}
\alias{anagram}
\alias{anagram-package}
\title{anagram: Comprehensive Anagram Detection and Analysis}
\description{
The "anagram" package offers a versatile set of tools for anagram detection, analysis, and visualization. Users can explore words from a dictionary that contain a specific subset of letters, with enriched output including frequency information for detailed analysis and visualization. The package also enables the determination of whether a set of words are anagrams of each other and facilitates the identification of anagrams within subsets of words. These functionalities empower users to delve into linguistic patterns, word relationships, and frequency distributions with ease and precision.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Advanced-R-Programming/assignment-3-package-completion-ajen0022}
  \item Report bugs at \url{https://github.com/Advanced-R-Programming/assignment-3-package-completion-ajen0022/issues}
}

}
\author{
\strong{Maintainer}: Amrita Jena \email{ajen0022@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
