% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_finder.R
\name{anagramfinder_function}
\alias{anagramfinder_function}
\title{Anagram Finder with Frequency Information}
\usage{
anagramfinder_function(word, dictionary)
}
\arguments{
\item{word}{The input word.}

\item{dictionary}{A data frame containing the dictionary of words.}
}
\value{
An S3 object of class 'anagram_output'.
}
\description{
This function creates an S3 object containing a list of words found in a dictionary with a specific subset of letters, along with their relative frequency. The output can be printed and plotted to visualize the relative frequencies of output words. A dictionary with frequency information is present for this extension.
}
\examples{
# Search for anagrams of the word "listen" in the dictionary
output <- anagramfinder_function("listen",dictionary)
# Print the output
print(output)
# Plot the output
plot(output)
}
