% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anagramsubset_function.R
\name{anagramsubset_function}
\alias{anagramsubset_function}
\title{Check anagram subset with minimum character length}
\usage{
anagramsubset_function(words, min_length = 1)
}
\arguments{
\item{words}{A character vector containing the words to be checked.}

\item{min_length}{An integer specifying the minimum length of combinations of characters to consider. Default is 1.}
}
\value{
TRUE if the subset contains anagrams of each other, FALSE otherwise.
}
\description{
This function checks whether a subset of words contains anagrams of each other,considering combinations of characters of a specified minimum length.
}
\examples{
# Example 1: Subset contains anagrams
 anagramsubset_function(c("tiger", "right", "tight"), 4)

# Example 2: Subset doesn't contain anagrams
anagramsubset_function(c("hello", "world", "hi"), 3)


}
