% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_anagrams.R
\name{are_anagrams}
\alias{are_anagrams}
\title{Test if a set of words are anagrams of each other}
\usage{
are_anagrams(words, subsets = FALSE)
}
\arguments{
\item{words}{A character vector containing the words to be tested.}

\item{subsets}{Logical indicating whether to check for anagrams only within subsets of words. Default is FALSE.}
}
\value{
TRUE if the words are all anagrams of each other, FALSE otherwise.
}
\description{
This function determines whether a set of words are anagrams of each other. Anagrams are words that contain the same letters but may be arranged differently. The function takes a character vector of words as input and checks if they all share the same set of letters, indicating that they are anagrams.
}
\examples{
# Example 1: checking if multiple words are anagrams
words <- c("actors", "costar", "castor")
are_anagrams(words)

# Example 2: Checking for anagrams within subsets
words <- c("alert", "alter", "hello", "world")
are_anagrams(words, subsets = TRUE)

}
