% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_finder.R
\name{most_common_word}
\alias{most_common_word}
\alias{most_common_word.anagram_output}
\alias{most_common_word.default}
\title{Find the most common word among the anagrams}
\usage{
most_common_word(x, ...)

\method{most_common_word}{anagram_output}(x, ...)

\method{most_common_word}{default}(x, ...)
}
\arguments{
\item{x}{An S3 object of class 'anagram_output' representing matches found among the anagrams of the input word.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
The most common word found among the anagrams.
}
\description{
The \code{most_common_word} function extracts the most common word from the matches found among the anagrams.
If a specific method is not defined for the class of the input object, a default method is called, which raises an error.

Method for extracting the most common word from an S3 object of class 'anagram_output'.

Default method for \code{most_common_word} function. Raises an error when no specific method is defined for the class of the input object.
}
\details{
This function extracts the most common word from the matches found among the anagrams of the input word. An anagram is a word that can be formed by rearranging the letters of the input word.
}
\examples{
# Generate matches among the words containing a subset of letters from the input word "when"
output <- anagramfinder_function("when", dictionary)
most_common_word(output)
}
